requirements.txt
fastapi
uvicorn
pyYAML
pymongo[srv]
pydantic
pytest
coverage